+++ lib/labels/entities/printer.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module LabelApp
+  class Printer < Dry::Struct
+    attribute :id, Types::Integer
+    attribute :printer_code, Types::String
+    attribute :printer_name, Types::String
+    attribute :printer_type, Types::String
+    attribute :pixels_per_mm, Types::Integer
+    attribute :printer_language, Types::String
+  end
+end